% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_bulk_data.R
\docType{data}
\name{bulkH1data}
\alias{bulkH1data}
\title{Example datasets for SCnorm}
\format{data matrix}
\usage{
data(bulkH1data)
}
\description{
H1 bulk data from paper, samples were multiplexed at 48 per lane.
}
\examples{
data(bulkH1data)
}
\keyword{datasets}
