% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCountDepth.R
\name{checkCountDepth}
\alias{checkCountDepth}
\title{Evaluation the count-depth relationship before (or after) normalizing the data.}
\usage{
checkCountDepth(Data, NormalizedData= NULL, Conditions = NULL, OutputName, SavePDF = TRUE, Tau = .5, FilterCellProportion = .10, 
                           FilterExpression = 0, NumExpressionGroups = 10, NCores=NULL, ditherCounts = FALSE)
}
\arguments{
\item{Data}{matrix of un-normalized expression counts. Rows are genes and columns are samples.}

\item{NormalizedData}{matrix of normalized expression counts. Rows are genes and columns are samples. Only input this if evaluating already normalized data.}

\item{Conditions}{vector of condition labels, this should correspond to the columns of the un-normalized expression matrix. If not provided data is assumed to come from same condition/batch.}

\item{OutputName}{specify the path and/or name of output files.}

\item{Tau}{value of quantile for the quantile regression used to estimate gene-specific slopes (default is median, Tau = .5 ).}

\item{SavePDF}{whether to automatically write and save the output plot as a PDF (default is TRUE).}

\item{FilterCellProportion}{the proportion of non-zero expression estimates required to include the genes into the evaluation. Default is .10.}

\item{FilterExpression}{exclude genes having median of non-zero expression below this threshold from count-depth plots.}

\item{NumExpressionGroups}{the number of groups to split the data into, groups are split into equally sized groups based on non-zero median expression.}

\item{NCores}{number of cores to use, default is detectCores() - 1.}

\item{ditherCounts}{whether to dither/jitter the counts, may be used for data with many ties, default is FALSE.}
}
\value{
outputs a plot.
}
\description{
Quantile regression is used to estimate the dependence of read counts on sequencing depth for every gene. If multiple conditions are provided, a separate plot is provided for each. Can be used to evaluate the extent of the count-depth relationship in the dataset or can be be used to evaluate data normalized by alternative methods.
}
\author{
Rhonda Bacher
}

