% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWithinFactor.R
\name{checkWithinFactor}
\alias{checkWithinFactor}
\title{Evaluation within sample factors in the the data.}
\usage{
checkWithinFactor(Data, withinSample=NULL, Conditions = NULL, 
    FilterExpression = 0, 
    NumExpressionGroups = 4)
}
\arguments{
\item{Data}{matrix of un-normalized expression counts. Rows are genes
and columns are samples.}

\item{withinSample}{a vector of gene-specific features.}

\item{Conditions}{vector of condition labels, this should correspond to 
 the columns of the un-normalized expression matrix. If provided the cells
will be colored by Condition instead of individually.}

\item{FilterExpression}{exclude genes having median of non-zero expression
below this threshold.}

\item{NumExpressionGroups}{the number of groups to split the within 
sample factor into, e.g genes will be split into equally sized groups
based on their GC content/Gene length/etc.}
}
\value{
produces a plot and returns the data the plot is based on.
}
\description{
This function can be used to evaluate the extent of 
  gene-specific biases in the data. If a bias exists, the plots provided 
  here will identify whether it affects cells equally or not. Correction 
  for such feature may be considered especially if the bias if different 
  between conditions (see SCnorm vignette for details).
}
\examples{
 
data(ExampleData)
Conditions = rep(c(1,2), each= 90) 
exampleFactor = runif(dim(ExampleData)[1], 0, 1)
#checkWithinFactor(Data = ExampleData,  
  #withinSample=exampleFactor, Conditions = Conditions)
}
\author{
Rhonda Bacher
}
