% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuncs.R
\name{correctWithin}
\alias{correctWithin}
\title{correctWithin}
\usage{
correctWithin(y, correctFactor)
}
\arguments{
\item{x}{gene to perform the regression on.}

\item{InputData}{list of data needed for the regression.}
}
\value{
within-cell normalized expression estimates
}
\description{
Perform the correction within each sample (See loess normalization in 
original publication Risso et al., 2011 (BMC Bioinformatics)).
Similar to function in EDAseq v2.8.0.
}
\details{
Performs within sample normalization.
}
