% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sim_data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example datasets for SCnorm}
\format{data matrix}
\usage{
data(ExampleData)
}
\description{
Data generated as in SIM I with K = 4
}
\examples{
data(ExampleData)
}
\keyword{datasets}

