% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleSimSCData.R
\docType{data}
\name{ExampleSimSCData}
\alias{ExampleSimSCData}
\title{Example datasets for SCnorm}
\format{data matrix}
\usage{
ExampleSimSCData
}
\description{
Data generated as in SIM I from the manuscript with K = 4.
}
\examples{
data(ExampleSimSCData)
}
\keyword{datasets}
