% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuncs.R
\name{getDens}
\alias{getDens}
\title{getDens}
\usage{
getDens(ExprGroups, byGroup, RETURN = c("Mode", "Height"))
}
\arguments{
\item{ExprGroups}{expression groups already split.}

\item{byGroup}{factor (usually slopes) to get density based on ExprGroups.}

\item{RETURN}{whether to return Mode or Height of density.}
}
\value{
list, length is equal to NumGroups
}
\description{
getDens
}
\details{
get density of slopes in different expression groups
}
