% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetK.R
\name{GetK}
\alias{GetK}
\title{Evaluate normalization using K slope groups}
\usage{
GetK(Data, SeqDepth, OrigData, Slopes, Name, PLOT = TRUE)
}
\arguments{
\item{Data}{matrix of normalized expression counts. Rows are genes and columns are samples.}

\item{SeqDepth}{vector of sequencing depths estimated as columns sums of un-normalized expression matrix.}

\item{OrigData}{matrix of un-normalized expression counts. Rows are genes and columns are samples.}

\item{Slopes}{vector of slopes estimated in the GetSlopes() function. Only used here to obtain the names of genes considered in the
normalization.}

\item{Name}{plot title}

\item{PLOT}{whether to output the evaluation plots.}
}
\value{
value of largest mode and (if PLOT = TRUE) a plot of the ten normalized slope densities.
}
\description{
Median quantile regresion is fit for each gene using the normalized gene expression values. 
A slope near zero indicate the sequencing depth effect has been successfully removed. 
Genes are divided into ten equally sized groups based on their non-zero median expression. Slope densities are plot for each 
group (if PLOT = TRUE) and estimated modes are calculated. If any of the ten group modes is larger than .1, the K is not sufficient
to normalize the data.
}
\author{
Rhonda Bacher
}

