% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSlopes.R
\name{GetSlopes}
\alias{GetSlopes}
\title{Estimate gene specfic count-depth relationships}
\usage{
GetSlopes(Data, SeqDepth, Tau)
}
\arguments{
\item{Data}{matrix of un-normalized expression counts. Rows are genes and columns are samples.}

\item{SeqDepth}{vector of sequencing depths estimated as columns sums of un-normalized expression matrix.}

\item{Tau}{value of quantile for the quantile regression used to estimate gene-specific slopes (default is median, Tau = .5 ).}
}
\value{
vector of estimated slopes.
}
\description{
This is the gene-specific fitting function, where a median (Tau = 
.5) quantile regression is fit for each gene. Only genes having at least 10 non-zero expression values are considered.
}
\author{
Rhonda Bacher
}

