% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFuncs.R
\name{quickreg}
\alias{quickreg}
\title{quickreg}
\usage{
quickreg(x, InputData)
}
\arguments{
\item{x}{gene to perform the regression on.}

\item{InputData}{list of data needed for the regression.}
}
\value{
gene slope.
}
\description{
Perform the single gene regressions using quantile regression.
}
\details{
Perform the single gene regressions using quantile regression.
}
