% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sc_data.R
\docType{data}
\name{scH1data}
\alias{scH1data}
\title{Example datasets for SCnorm}
\format{data matrix}
\usage{
data(scH1data)
}
\description{
H1 single cell data from paper, same cells sequenced to an average of 1 million reads per cell and 4 million reads per cell.
}
\examples{
data(scH1data)
}
\keyword{datasets}
