% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCnorm.R
\name{SCnorm}
\alias{SCnorm}
\title{SCnorm}
\usage{
SCnorm(Data, Conditions, K, OutputName = "SCnorm", PLOT = TRUE, PropToUse = .25, outlierCheck = TRUE, Tau = .5, reportSF = FALSE)
}
\arguments{
\item{Data}{matrix of un-normalized expression counts. Rows are genes and columns are samples.}

\item{Conditions}{vector of condition labels, this should correspond to the columns of the un-normalized expression matrix.}

\item{OutputName}{specify the path and/or name of output files.}

\item{PLOT}{whether to save all output the evaluation plots while determining the optimal K.}

\item{PropToUse}{proportion of genes closest to the slope mode used for the group fitting, default is set at .25. This number mainly affects speed.}

\item{outlierCheck}{cells/samples with relatively small sequencing depths may recieve very small scaling factors, to ensure these 
small scaling factors do not create outliers, genes with potential outliers are first flagged and then if necessary, their normalized expression 
values are corrected. First, gene expression counts in cells/samples having the three smallest sequencing depths are flagged if they contain
expression counts larger than 3 times the predicted counts (from the group regression). For each gene, if the flagged values
are the largest normalized expression value, then the count is corrected to be the median of the non-zero normalized values (default = 
FALSE)}

\item{Tau}{value of quantile for the quantile regression used to estimate gene-specific slopes (default is median, Tau = .5 ).}

\item{reportSF}{whether to provide a matrix of scaling counts in the output (default = FALSE).}

\item{K}{the number of groups for normalizing. If left unspecified, an evaluation procedure will determine the optimal value of K
(recommended). If you're sure about specifiyng K, then a vector equal to the number of conditions may be used.}

\item{NCores}{number of cores to use, default is detectCores() - 1.}

\item{FilterCellNumber}{the number of non-zero expression estimate required to include the genes into the SCnorm fitting (default = 10). THe initial 
grouping fits a quantile regression to each gene, making this value too low gives unstable fits.}
}
\value{
List containing matrix of normalized expression (and optionally a matrix of size factors if reportSF = TRUE ).
}
\description{
Quantile regression is used to estimate the dependence of read counts on sequencing depth for every gene. Genes
with similar dependence are then grouped, and a second quantile regression is used to estimate scale factors within each 
group. 
Within-group adjustment for sequencing depth is then performed using the estimated scale factors to provide normalized 
estimates of expression. If multiple conditions are provided, normalization is performed within condition and then
normalized estimates are scaled between conditions.
}
\author{
Rhonda Bacher
}

