% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuncs.R
\name{splitGroups}
\alias{splitGroups}
\title{splitGroups}
\usage{
splitGroups(DATA, NumGroups = 10)
}
\arguments{
\item{DATA}{vector to be splot.}

\item{NumGroups}{number of groups}
}
\value{
list, length is equal to NumGroups
}
\description{
splitGroups
}
\details{
helper function to get split a vector into a specified number of groups
}
