% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialEvalPlot.R
\name{initialEvalPlot}
\alias{initialEvalPlot}
\title{Evaluate data before normalization.}
\usage{
initialEvalPlot(Data, SeqDepth, Slopes, Name, NumExpressionGroups = 10)
}
\arguments{
\item{SeqDepth}{vector of sequencing depths estimated as columns sums of un-normalized expression matrix.}

\item{Slopes}{vector of slopes estimated in the GetSlopes() function.}

\item{Name}{plot title}

\item{NumExpressionGroups}{number of (non-zero median) expression groups to split the slope densities into.}

\item{BeforeNorm}{whether data are already normalized or not.}

\item{Data}{matrix of normalized expression counts. Rows are genes and columns are samples.}
}
\value{
a plot of the un-normalized slope densities.
}
\description{
Genes are divided into NumExpressionGroups = 10 equally sized groups based on their non-zero median expression. Slope densities are plot for each group.
}
\author{
Rhonda Bacher
}

