% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormWrap.R
\name{Normalize}
\alias{Normalize}
\title{Iteratively fit group regression and evaluate to choose optimal K}
\usage{
Normalize(Data, SeqDepth, Slopes, CondNum, PLOT = TRUE,
  OutputName = OutputName, PropToUse, Tau, NCores, Thresh, ditherCounts)
}
\value{
matrix of normalized and scaled expression values for all conditions. If PLOT = TRUE then the evaluation plots are 
output for each attempted value of K.
}
\description{
This function iteratively normalizes using K groups and then evaluates whether K is sufficient. If the maximum mode received 
from the GetK() function is larger than .1, K is increased to K + 1. Uses params sent from SCnorm.
}
\author{
Rhonda Bacher
}

