% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormWrap.R
\name{Normalize}
\alias{Normalize}
\title{Iteratively fit group regression and evaluate to choose optimal K}
\usage{
Normalize(Data, SeqDepth, Slopes, CondNum, 
   OutputName = OutputName, PropToUse, Tau, NCores, Thresh, ditherCounts)
}
\arguments{
\item{Data}{matrix of un-normalized expression counts. Rows are genes and
columns are samples.}

\item{SeqDepth}{sequencing depth for each cell/sample.}

\item{Slopes}{per gene estimates of the count-depth relationship.}

\item{CondNum}{name of group being normalized, just for printing messages.}

\item{OutputName}{specify the path and/or name of output files.}

\item{PropToUse}{proportion of genes closest to the slope mode used for
the group fitting, default is set at .25. This number #' mainly affects
speed.}

\item{Tau}{value of quantile for the quantile regression used to estimate
gene-specific slopes (default is median, Tau = .5 ).}

\item{NCores}{number of cores to use, default is detectCores() - 1.}

\item{Thresh}{threshold to use in evaluating the sufficiency of K, default
is .1.}

\item{ditherCounts}{whether to dither/jitter the counts, may be used for
data with many ties, default is FALSE.}
}
\value{
matrix of normalized and scaled expression values for all conditions
   and the evaluation plots are 
output for each attempted value of K.
}
\description{
This function iteratively normalizes using K groups and then
   evaluates whether K is sufficient. If the maximum mode received 
from the GetK() function is larger than .1, K is increased to K + 1. Uses
   params sent from SCnorm.
}
\author{
Rhonda Bacher
}
