% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFuncs.R
\name{redobox}
\alias{redobox}
\title{redobox}
\usage{
redobox(DATA, smallc)
}
\arguments{
\item{DATA}{data set to.}

\item{smallc, }{what value to ignore, typically is zero.}
}
\value{
the dataset has been logged with values below smallc masked.
}
\description{
redobox
}
\details{
Function to log data and turn zeros to NA to mask/ignore in 
 functions.
}
