% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCnorm_function.R
\name{SCnorm_fit}
\alias{SCnorm_fit}
\title{Fit group quantile regression for K groups}
\usage{
SCnorm_fit(Data, SeqDepth, Slopes, K, PropToUse = .25, outlierCheck = TRUE, Tau = .5)
}
\arguments{
\item{Data}{matrix of un-normalized expression counts. Rows are genes and columns are samples.}

\item{SeqDepth}{vector of sequencing depths estimated as columns sums of un-normalized expression matrix.}

\item{Slopes}{vector of slopes estimates from GetSlopes().}

\item{K}{number of slope groups in data, this will be estimated iteratively.}

\item{PropToUse}{proportion of genes closest to the slope mode used for the group fitting, default is set at .25. This number mainly affects speed.}

\item{outlierCheck}{cells/samples with relatively small sequencing depths may recieve very small scaling factors, to ensure these 
small scaling factors do not create outliers, genes with potential outliers are first flagged and then if necessary, their normalized expression 
values are corrected. First, gene expression counts in cells/samples having the three smallest sequencing depths are flagged if they contain
expression counts larger than 3 times the predicted counts (from the group regression). For each gene, if the flagged values
are the largest normalized expression value, then the count is corrected to be the median of the non-zero normalized values (default = 
FALSE)}

\item{Tau}{value of quantile for the quantile regression used to estimate gene-specific slopes (default is median, Tau = .5 ).}
}
\value{
normalized expression matrix and matrix of scaling factors.
}
\description{
For each group K, a quantile regression is fit over all genes (PropToUse) for a grid of possible degree's d and quantile's tau. 
For each value of tau and d, the predicted expression values are obtained and regressed against the original sequencing depths.  The optimal tau and d combination is chosen as that closest to the mode of the gene slopes.
}
\author{
Rhonda Bacher
}

